Fixing CSS for Windows browser
Google Chrome in Windows was picking up .woff format instead of svg.  This caused aliasing on the font.  Now that svg has priority over woff (in the CSS), it renders correctly in the browser.

In Windows 7, font looks clean across Chrome, Firefox and IE9.

Also re-exported font for fixing a little bit of aliasing in IE8.

Signed-off-by: Gregory Loucas <loucasgp@gmail.com>